Add files via upload
This is a notebook that is used to transform a query in BigQuery into a `sample.jsonl` file as well as a file for each id holding each text, so that this combination can be used for AutoML categorical NLP predictions.  

this notebook could and should be transformed into a single python file so that one doesn't have to do this from a notebook 
The difficulty encountered is simply that I'm using the cell magic `%%bigquery` command in order to transform the query into a pandas data frame and replacing this cell magic with pure python client functions has been buggy so far.